BackwardAlgorithm

Documentation

type Beta p = p

Label Type for the Backward Variable(Beta)

type BetaMap s p = Map s p

beta_T :: Num p => st -> Beta p

The initialization step for the backward algorithm

betas_T :: (Num p, Foldable t, Ord s) => HMM s o p t -> BetaMap s p

beta_t :: (Foldable t, Num p, Ord s) => HMM s o p t -> BetaMap s p -> o -> s -> Beta p

betas_t :: (Foldable t, Num p, Ord s) => HMM s o p t -> o -> BetaMap s p -> BetaMap s p

calculation of one time unit's beta variables

betas :: (Foldable t, Num p, Ord s) => HMM s o p t -> ObservationSeq o -> BetaMap s p

calculates the beta values for the first observation in a sequence

Very sure this is not a very efficient implementation