BaumWelch

Documentation

monitoredTraining :: (Traversable t, Fractional p, Show s, Show o, Show p, Ord s, Ord o) => ObservationSeq o -> Int -> Int -> HMM s o p t -> IO (HMM s o p t)

Estimation Step

type EstimationStep s o p t m = ReaderT (TimeMap (TimeData s o p), HMM s o p t, p) m

getsTM :: Monad m => (TimeMap (TimeData s o p) -> a) -> EstimationStep s o p t m a

getsTD :: Monad m => Int -> (TimeData s o p -> a) -> EstimationStep s o p t m a

getAlpha :: (Monad m, Num p, Ord s) => Int -> s -> EstimationStep s o p t m (Alpha p)

getBeta :: (Monad m, Num p, Ord s) => Int -> s -> EstimationStep s o p t m (Beta p)

getObs :: Monad m => Int -> EstimationStep s o p t m o

getsHMM :: Monad m => (HMM s o p t -> a) -> EstimationStep s o p t m a

getA_ij :: Monad m => s -> s -> EstimationStep s o p t m p

getB_j :: Monad m => s -> EstimationStep s o p t m (o -> p)

getLikelihood :: Monad m => EstimationStep s o p t m p

calcXi :: (Monad m, Fractional p, Ord s) => Int -> s -> s -> EstimationStep s o p t m p

calcGamma' :: (Traversable t, Monad m, Num p, Fractional p, Foldable t, Ord s) => Int -> s -> EstimationStep s o p t m p

Note: Source of inefficiency because this recalculations all Xi a bunch of times

calcGamma :: (Traversable t, Monad m, Num p, Fractional p, Foldable t, Ord s) => Int -> s -> EstimationStep s o p t m p

estData_t :: (Traversable t, Monad m, Ord s, Num p, Fractional p) => Int -> EstimationStep s o p t m (EstimationData s o p)

estData :: (Traversable t, Monad m, Ord s, Num p, Fractional p) => EstimationStep s o p t m (TimeMap (EstimationData s o p))

Maximization Step

We can reuse the estimation monad stack for this section

type MaximizationStep s o p t = ReaderT (TimeMap (EstimationData s o p), HMM s o p t)

getGamma :: Monad m => Int -> MaximizationStep s o p t m (Map s p)

getXi :: Monad m => Int -> MaximizationStep s o p t m (Map s (Map s p))

getObsActualAt :: Monad m => Int -> MaximizationStep s o p t m (Int, o)

getsHMM_M :: Monad m => (HMM s o p t -> a) -> MaximizationStep s o p t m a

pi' :: (Monad m, Traversable t, Fractional p, Ord s) => MaximizationStep s o p t m (s -> p)

foreachStateTransition :: (Monad m, Traversable t, Ord s) => (s -> s -> MaximizationStep s o p t m a) -> MaximizationStep s o p t m (Map s (Map s a))

complete some monadic action for each state transition

foreachStateAndObs :: (Monad m, Traversable t, Ord s, Ord o) => (s -> o -> MaximizationStep s o p t m a) -> MaximizationStep s o p t m (Map s (Map o a))

mapToTwoArgFunc :: (Ord k1, Ord k2) => Map k1 (Map k2 v) -> k1 -> k2 -> v

alpha' :: (Monad m, Traversable t, Fractional p, Ord s) => MaximizationStep s o p t m (s -> s -> p)

beta' :: (Monad m, Traversable t, Fractional p, Ord s, Ord o, Eq o) => MaximizationStep s o p t m (s -> o -> p)

normalizeDist :: (Fractional p, Ord k) => Map k p -> Map k p

maximization :: (Monad m, Traversable t, Fractional p, Ord s, Ord o, Eq o) => MaximizationStep s o p t m (HMM s o p t)

mapToFunc :: Ord k => Map k a -> k -> a

Bringing it All Together

reestimate :: (Traversable t, Ord s, Num p, Fractional p, Ord o, Eq o) => ObservationSeq o -> HMM s o p t -> HMM s o p t

initData :: (Traversable t, Num p, Ord s) => HMM s o p t -> ObservationSeq o -> (TimeMap (TimeData s o p), HMM s o p t, p)

initialTimeData :: (Traversable t, Num p, Ord s) => HMM s o p t -> ObservationSeq o -> TimeMap (TimeData s o p)

computeAlphas :: (Traversable t, Num p, Ord s) => HMM s o p t -> ObservationSeq o -> TimeMap (AlphaMap s p)

computeBetas :: (Traversable t, Num p, Ord s) => HMM s o p t -> ObservationSeq o -> TimeMap (BetaMap s p)

computeLikelihood :: Num p => TimeMap (TimeData s o p) -> p

Time-Tagged Data

type Time = Int

type TimeMap a = IntMap a

data TimeData s o p

Time ordered forward/backward variables with their respective observations

Constructors

TimeData

Fields

alpha :: AlphaMap s p

beta :: BetaMap s p

t_obs :: o

data EstimationData s o p

Time ordered estimation-step data

Constructors

EstData

Fields

gamma :: Map s p

xi :: Map s (Map s p)

e_obs :: o

mkTimeData :: TimeMap (AlphaMap s p) -> TimeMap (BetaMap s p) -> ObservationSeq o -> TimeMap (TimeData s o p)

aggregation of TimeMaps AlphaMap and BetaMap with their observation sequence