ForwardAlgorithm

Likelihood (Problem 1)

likelihood :: (Traversable t, Functor t, Num p, Foldable seq, Ord s) => HMM s o p t -> ObservationSeq o seq -> p

the probability of an observation sequence occurring given a model

this is problem 1, or the likelihood problem, in most HMM texts

Calculation of the Forward Variable(Alpha)

type Alpha p = p

type AlphaMap s p = Map s p

alphas_T :: (Traversable t, Functor t, Num p, Foldable seq, Ord s) => HMM s o p t -> ObservationSeq o seq -> AlphaMap s p

The alpha calculation made for an entire observation sequence This is summed over to achieve the probability of a sequence occurring

alphas_t :: (Num p, Ord s, Foldable t) => AlphaMap s p -> HMM s o p t -> o -> AlphaMap s p

the alphas for observation o_t with respect to the alphas from the previous observation.

alphas_1 :: (Num p, Ord s, Foldable t) => HMM s o p t -> o -> AlphaMap s p

The alphas for observation o_1 given a model

alphas' :: (Num p, Ord s, Foldable t) => (StateInfo s o p -> o -> Alpha p) -> HMM s o p t -> o -> AlphaMap s p

the abstract calculation for the alphas of an observation o with respect to a model and observation. The first argument is a function that calculates the alpha value for a given state s_i given its StateInfo view and the observation. \ This is used to provide an abstraction so that addition data can be curried into the function before alpha's is called.

alpha1 :: Num p => StateInfo s o p -> o -> Alpha p

the calculation of a forward variable with respect to the first observation in a sequence

alpha_t :: Num p => AlphaMap s p -> StateInfo s o p -> o -> Alpha p

the calculation for a forward variable for any (non-initial) observation with respect to the previous observation's forward variables, the current state's(s_j) state info

pInto :: Num p => StateInfo s o p -> AlphaMap s p -> p

The probability of transitioning into state s_j from all other states given the forward variable from the previous observation