HMM

Documentation

type Probability = CReal

Hidden Markov Models

data HMM s o p t

The definition of a hidden markov model s := data type of states o := data type of observations p := data type of probabilities t := a traversable functor to contain the states and observations

Constructors

HMM

Fields

states :: t s

observations :: t o

initialProbability :: s -> p

stateTransitions :: s -> s -> p

observationProbability :: s -> o -> p

stateCount :: Traversable t => HMM s o p t -> Int

observationCount :: Traversable t => HMM s o p t -> Int

data StateInfo s o p

Constructors

StateInfo

Fields

pi_s :: p

p_inbound :: s -> p

p_outbound :: s -> p

p_obs :: o -> p

stateInfo :: HMM s o p t -> s -> StateInfo s o p

overStateSpace :: (Foldable t, Ord s) => HMM s o p t -> (s -> a) -> Map s a

data ObservationSeq o seq

Constructors

ObsSeq

Fields

first :: o

rest :: seq o

fromList :: [o] -> ObservationSeq o []

data StateSeq o seq

Constructors

StateSeq (seq o)

pIntoG :: Num p => (p -> p -> p) -> StateInfo s o p -> Map s p -> p

a function applied to the probabilities of transition(a_ij * alpha_i) from s_i to s_j

accumMultiply :: Num p => (p -> p -> p) -> StateInfo s o p -> s -> p -> p -> p

Helper functions for Monads

HMMComp - Stateful actions within an HMM environment(reader)

Towards a more declarative algorithm