Pretty

Documentation

showHMM :: (Foldable t, Ord s, Ord o, Show p, Show s, Show o) => HMM s o p t -> String

showPIs :: (Foldable t, Ord s, Show p, Show s) => HMM s o p t -> Box

showAlphas :: (Foldable t, Ord s, Show p, Show s) => HMM s o p t -> Box

showBetas :: (Foldable t, Ord s, Ord o, Show p, Show s, Show o) => HMM s o p t -> Box

collapseDoubleMap :: (Ord k1, Ord k2) => Map k1 (Map k2 v) -> Map (k1, k2) v

showEmission :: (Show k, Show e) => (k, e) -> String

showTransition :: (Show k1, Show k2) => (k1, k2) -> String

prettyMap' :: Show v => (k -> String) -> Map k v -> Box

prettyMap :: (Show v, Show k) => Map k v -> Box

prettyPercent :: PrintfArg a => a -> String