Viterbi

Decoding (Problem 2)

decode :: (Traversable t, Functor t, Num p, Ord p, Foldable seq, Ord s) => HMM s o p t -> ObservationSeq o seq -> V s p

Viterbi Variable

data V s p

The viterbi variable aggregate which includes the probability and (most likely) previous state for backtracing

Constructors

V (Seq s) p

Vi p

vNil :: Num p => V s p

probv :: V s p -> p

seqV :: V s p -> Seq s

chain :: s -> p -> V s p -> V s p

type ViterbiMap s p = Map s (V s p)

Viterbi Variable Calculations

viterbis_T :: (Foldable seq, Num p, Ord p, Foldable t, Ord s) => HMM s o p t -> ObservationSeq o seq -> ViterbiMap s p

viterbis_1 :: (Num p, Ord s, Foldable t) => HMM s o p t -> o -> ViterbiMap s p

viterbi variable for the first observation(O_1) in a sequence

viterbis_t :: (Num p, Ord p, Foldable t, Ord s) => ViterbiMap s p -> HMM s o p t -> o -> ViterbiMap s p

viterbi_t :: (Num p, Ord p) => ViterbiMap s p -> StateInfo s o p -> o -> V s p

Utility Functions

argmax :: (Ord a, Foldable t, Functor t) => (b -> a) -> t b -> (a, b)

calculations the maximal value and its argument when a function f is applied to a foldable functor

pIntoV :: (Num p, Ord p) => StateInfo s o p -> Map s (V s p) -> V s p

Viterbi Algorithm: Finding the most likely state sequence given an observation sequence (Problem 2)